+++ spec/beautydate_api/business_spec.rb
@@ -7,7 +7,7 @@ describe BeautydateApi::Business do
     @response = JSON.load(File.read('./spec/fixtures/business.json'))
   end
 
-  before { BeautydateApi::APIRequest.stub(:request).with('GET', described_class.url(id_or_uuid)) { @response } }
+  before { BeautydateApi::Core::Request.stub(:request).with('GET', described_class.url(id_or_uuid)) { @response } }
 
 
   describe '#refresh' do
@@ -27,7 +27,7 @@ describe BeautydateApi::Business do
 
     context 'when neither id nor uuid are present' do
       let(:id_or_uuid) { nil }
-      it { expect { subject.refresh }.to raise_error(BeautydateApi::APIResource::UnkownIdentifierError) }
+      it { expect { subject.refresh }.to raise_error(BeautydateApi::Core::Resource::UnkownIdentifierError) }
     end
   end
 end
